fix(cli): deployer should wait for prereq txs
