+++ packages/cli/src/deploy/configToModules.ts
@@ -5,10 +5,8 @@ import { SchemaAbiType, SchemaAbiTypeToPrimitiveType } from "@latticexyz/schema-
 import { bytesToHex } from "viem";
 import { createPrepareDeploy } from "./createPrepareDeploy";
 import { World } from "@latticexyz/world";
-import { getContractArtifact } from "../utils/getContractArtifact";
 import { importContractArtifact } from "../utils/importContractArtifact";
 import { resolveWithContext } from "@latticexyz/world/internal";
-import metadataModule from "@latticexyz/world-module-metadata/out/MetadataModule.sol/MetadataModule.json" assert { type: "json" };
 
 /** Please don't add to this list! These are kept for backwards compatibility and assumes the downstream project has this module installed as a dependency. */
 const knownModuleArtifacts = {
@@ -19,28 +17,11 @@ const knownModuleArtifacts = {
     "@latticexyz/world-modules/out/Unstable_CallWithSignatureModule.sol/Unstable_CallWithSignatureModule.json",
 };
 
-const metadataModuleArtifact = getContractArtifact(metadataModule);
-
 export async function configToModules<config extends World>(
   config: config,
   // TODO: remove/replace `forgeOutDir`
   forgeOutDir: string,
 ): Promise<readonly Module[]> {
-  const defaultModules: Module[] = [
-    // TODO: replace metadata install here with custom logic inside `ensureModules` or an `ensureDefaultModules` to check
-    //       if metadata namespace exists, if we own it, and if so transfer ownership to the module before reinstalling
-    //       (https://github.com/latticexyz/mud/issues/3035)
-    {
-      optional: true,
-      name: "MetadataModule",
-      installAsRoot: false,
-      installData: "0x",
-      prepareDeploy: createPrepareDeploy(metadataModuleArtifact.bytecode, metadataModuleArtifact.placeholders),
-      deployedBytecodeSize: metadataModuleArtifact.deployedBytecodeSize,
-      abi: metadataModuleArtifact.abi,
-    },
-  ];
-
   const modules = await Promise.all(
     config.modules.map(async (mod): Promise<Module> => {
       let artifactPath = mod.artifactPath;
@@ -98,5 +79,5 @@ export async function configToModules<config extends World>(
     }),
   );
 
-  return [...defaultModules, ...modules];
+  return modules;
 }
